requirements.txt
Adyen==7.1.1
asgiref==3.5.2
attrs==23.1.0
base58check==1.0.2
bech32==1.2.0
blake256==0.1.1
btcpay-python==1.3.0
cairocffi==1.6.1
cbor==1.0.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.3
django-admin-confirm==1.0.0
django-appconf==1.0.5
django-htmx==1.13.0
django-imagekit==4.1.0
django-multiselectfield==0.1.12
django-nested-admin==4.0.2
django-object-actions==4.1.0
django-simple-history==3.3.0
django-widget-tweaks==1.4.12
ecdsa==0.18.0
groestlcoin-hash2==1.1.1
idna==3.4
lxml==4.9.3
pilkit==2.0
Pillow==9.3.0
pycparser==2.21
pysha3==1.0.2
python-monkey-business==1.0.0
pytz==2023.3.post1
reportlab==4.0.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
urllib3==1.26.13
webencodings==0.5.1
zope.interface==6.0
